+++ components/QuestionAdd/__tests__/index.spec.tsx
@@ -89,6 +89,14 @@ describe("QuestionAdd", () => {
     const mockTemplateId = 123;
     const mockUseParams = useParams as jest.Mock;
 
+    // Mock window.tinymce
+    window.tinymce = {
+      init: jest.fn(),
+      remove: jest.fn(),
+    };
+
+
+
     // Mock the return value of useParams
     mockUseParams.mockReturnValue({ templateId: `${mockTemplateId}` });
 
@@ -123,15 +131,15 @@ describe("QuestionAdd", () => {
     expect(editOptionsTab).toBeInTheDocument();
     const editLogicTab = screen.getByRole('tab', { name: 'tabs.logic' });
     expect(editLogicTab).toBeInTheDocument();
-    const questionTypeLabel = screen.getByRole('textbox', { name: 'labels.type' });
+    const questionTypeLabel = screen.getByLabelText(/labels.type/i);
     expect(questionTypeLabel).toBeInTheDocument();
-    const questionTextLabel = screen.getByRole('textbox', { name: 'labels.questionText' });
+    const questionTextLabel = screen.getByText(/labels.questionText/i);
     expect(questionTextLabel).toBeInTheDocument();
-    const questionRequirementTextLabel = screen.getByRole('textbox', { name: 'labels.requirementText' });
+    const questionRequirementTextLabel = screen.getByText(/labels.requirementText/i);
     expect(questionRequirementTextLabel).toBeInTheDocument();
-    const questionGuidanceTextLabel = screen.getByRole('textbox', { name: 'labels.guidanceText' });
+    const questionGuidanceTextLabel = screen.getByText(/labels.guidanceText/i);
     expect(questionGuidanceTextLabel).toBeInTheDocument();
-    const questionSampleTextLabel = screen.getByRole('textbox', { name: 'labels.sampleText' });
+    const questionSampleTextLabel = screen.getByText(/labels.sampleText/i);
     expect(questionSampleTextLabel).toBeInTheDocument();
     const sidebarHeading = screen.getByRole('heading', { level: 2 });
     expect(sidebarHeading).toHaveTextContent('headings.preview');