+++ components/QuestionView/index.tsx
@@ -25,7 +25,7 @@ import {
 } from "@/components/Card/card";
 
 
-import { DmpEditor } from "@/components/Editor";
+import TinyMCEEditor from '@/components/TinyMCEEditor';
 import { Button } from "react-aria-components";
 
 import styles from './QuestionView.module.scss';
@@ -104,7 +104,8 @@ const QuestionView: React.FC<QuestionViewProps> = ({
           <CardHeading>{question?.questionText}</CardHeading>
           <CardBody data-testid="card-body">
             {(questionType == 'Text Area') && (
-              <DmpEditor
+              <TinyMCEEditor
+                id="question-text-editor"
                 content={question?.useSampleTextAsDefault ? question.sampleText as string : ''}
                 setContent={() => { }} // Pass an empty function
               />