+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - QuestionPreview component [#224]
 
 ### Updated
+- Switched out all occurrences of Remirror/DMPEditor rich text editor with TinyMCE [#496]
 - Convert all references to the old `Contributor` to `Member` and `Project/PlanFunder` to `Project/PlanFunding` to mimic language used in UI
 - Added the `Account Profile` link on the home page for the demo and fixed the success message when updating profile [#508] [#512]
 - Updated the "add new section" page to properly split our the affiliation's sections from best practice sections [#451]